The game changer (aka new FFTLib)
upgrade to latest ArduinoFFT Lib, which allows to use float instead of double.
- Processing time is a lot shorter - new dev version: 3-5ms; old release: 20-25ms
This will enable us to implement several additions, like overlapping FFT (doubles the rate of FFTResults) and new onset detection based on "beatroot" algorithm.

As a side-effect, RAM usage is reduced.

+++ wled00/FX.cpp
@@ -4281,9 +4281,9 @@ extern float sampleReal;			                 // "sample" as float, to provide bit
 extern float multAgc;                          // sampleReal * multAgc = sampleAgc. Our multiplier
 
 // FFT based variables
-extern double FFT_MajorPeak;
-extern double FFT_Magnitude;
-extern double fftBin[];                         // raw FFT data
+extern float FFT_MajorPeak;
+extern float FFT_Magnitude;
+extern float fftBin[];                         // raw FFT data
 extern int fftResult[];                         // summary of bins array. 16 summary bins.
 extern float fftAvg[];
